JIRA Task #134
updated the cypress test for correlation.cy.ts and linear.cy.ts
Created userManual.cy.ts to test the usermanual page to make sure all parts are visible and the routerlinks work.

correlation and linear regression tests accurately check to see if the legends work for the chart plugins to make sure they work. It had to be mroe manual as it is not using html elements but the plugin. Screenshots using Cypress allowed me to make sure the test positions are accurate and working.